Exclude direnv tests from Windows
